Getting taskflow running with setup_cluster move.
Trying to get the taskflows working with the setup_cluster method
now residing in cumulus. Moved setup_cluster to taskflow.core but
once I try and import it into taskflow I get the similar import
errors.
